+++ src/index.js
@@ -1,10 +1,7 @@
-import React from 'react'
 import express from 'express'
 import fs from 'fs'
 import path from 'path'
 import 'regenerator-runtime/runtime'
-import { createElement, pdf, PDFRenderer } from '@react-pdf/core'
-import { createApolloFetch } from 'apollo-fetch'
 
 if (process.env.NODE_ENV !== 'production') {
   require('dotenv').config()
@@ -12,69 +9,10 @@ if (process.env.NODE_ENV !== 'production') {
 
 const PORT = process.env.PORT || 3007
 
-const Document = require('./components/Document').default
-
-const query = `
-  query getDocument($path: String!) {
-    article: document(path: $path) {
-      id
-      content
-      meta {
-        template
-        path
-        title
-        description
-        image
-        color
-        format {
-          meta {
-            path
-            title
-            color
-            kind
-          }
-        }
-      }
-    }
-  }
-`
+const { renderDocument } = require('./render')
+const { getDocument } = require('./fetch')
 
 const server = express()
-
-const render = async (article, query, response) => {
-  const container = createElement('ROOT')
-  const node = PDFRenderer.createContainer(container)
-
-  const format = article.meta.format || {}
-  const formatMeta = format.meta || {}
-  const formatTitle = formatMeta.title
-  const formatColor = formatMeta.color
-
-  PDFRenderer.updateContainer(
-    <Document article={article} options={{
-      formatTitle,
-      formatColor,
-      images: query.images !== '0'
-    }} />,
-    node,
-    null
-  )
-
-  response.set('Content-Type', 'application/pdf')
-  if (query.download) {
-    const fileName = article.meta.path
-      .split('/')
-      .filter(Boolean)
-      .join('-')
-    response.set(
-      'Content-Disposition',
-      `attachment; filename="${fileName}.pdf"`
-    )
-  }
-
-  const output = await pdf(container).toBuffer()
-  output.pipe(response)
-}
 
 const stripDotPdf = path => path.replace(/\.pdf$/, '')
 
@@ -90,24 +28,19 @@ server.get('/fixtures/:path', (req, res) => {
   const api = JSON.parse(
     fs.readFileSync(fixturePath, 'utf8')
   )
-  render(api.data.document, req.query, res)
+  renderDocument(api.data.document, req.query, res)
 })
 
 server.get('/:path*', async (req, res) => {
-  const variables = {
+  const api = await getDocument({
     path: stripDotPdf(req.path)
-  }
-
-  const apolloFetch = createApolloFetch({
-    uri: process.env.API_URL
   })
-  const api = await apolloFetch({ query, variables })
 
   if (!api.data.article) {
     res.status(404).end('No Article Found')
     return
   }
-  render(api.data.article, req.query, res)
+  renderDocument(api.data.article, req.query, res)
 })
 
 server.listen(PORT, err => {
